requirements.txt
# Core Dependencies
fastapi
uvicorn==0.27.0   # Updated to a newer version
python-dotenv==1.0.0

# Data Handling
pandas==2.1.1
numpy==1.26.0

# Machine Learning
transformers==4.34.0

# API Clients
requests==2.31.0
openai


# Testing
pytest==7.4.2
pytest-cov==4.1.0



# Optional (for deployment)
gunicorn==21.2.0